glpi_pdf_project/glpi_app/requirements.txt
requests
python-dotenv
unstructured
langchain
langchain-community
chromadb
pydantic
typing-extensions
fastapi
uvicorn
boto3
reportlab
openai
sentence-transformers
